ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
PyMuPDF==1.24.5
python-docx==1.1.2
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
regex==2024.5.15
pydantic==2.7.4